[MP-1534] Update concurrency wrapper
This PR: Updates the concurrency wrapper with the 15.0.0 version of iOS Marigold SDK

Task: https://sailthru.atlassian.net/browse/MP-1534
