+++ README.md
@@ -1,6 +1,6 @@
-# SailthruMobileSwift
+# MarigoldSwift
 
-This package acts as a wrapper around the [Sailthru Mobile iOS SDK](https://github.com/sailthru/sailthru-mobile-ios-sdk) to allow the asynchronous functionality to be accessed using Swift Concurrency.
+This package acts as a wrapper around the [Marigold iOS SDK](https://github.com/sailthru/sailthru-mobile-ios-sdk) to allow the asynchronous functionality to be accessed using Swift Concurrency.
 
 It can be added to a project as a Swift Package and will include the base SDK.
